choicemate-flask-api/requirements.txt
Flask
pandas
numpy
beautifulsoup4
python-dotenv
google-generativeai
